+++ leetcode/200_Number_of_Islands.h
@@ -3,7 +3,7 @@
 class L200 : public icode {
 public:
     void dfs(vector<vector<char>>& grid, int i, int j) {
-        if (i < 0 || j < 0 || i >= grid.size() || j >= grid[0].size() || grid[i][j]=='0')
+        if (i < 0 || j < 0 || i >= (int)grid.size() || j >= (int)grid[0].size() || grid[i][j]=='0')
             return;
         grid[i][j] = '0';
         dfs(grid, i, j+1);
